refactor(pricing): consolidate pricing into @genfeedai/pricing single source of truth (#486)
Create packages/pricing as the canonical home for all pricing code:
provider cost tables (from @genfeedai/core), plans/credits/BYOK
constants (from @genfeedai/helpers), and getPricingConfig (from
@genfeedai/config). The three old locations become pure named
re-exports, so every existing import path keeps working.

Wire the Joi defaults in stripe.schema.ts to the canonical
BYOK_FEE_PERCENTAGE / BYOK_FREE_THRESHOLD_CREDITS constants instead of
duplicated literals, and drop the dead || 5 / || 500 fallbacks in
byok-billing.service.ts (Joi .default() guarantees values at startup;
the fallbacks also silently coerced an explicit env value of 0 back to
5/500).

Add @genfeedai/pricing path mappings and vitest aliases everywhere
@genfeedai/core or @genfeedai/helpers are compiled from source.
