+++ apps/app/tsconfig.json
@@ -100,6 +100,8 @@
       "@genfeedai/fonts": ["../../packages/fonts/index.ts"],
       "@genfeedai/helpers": ["../../packages/helpers/src/index.ts"],
       "@genfeedai/helpers/*": ["../../packages/helpers/src/*"],
+      "@genfeedai/pricing": ["../../packages/pricing/src/index.ts"],
+      "@genfeedai/pricing/*": ["../../packages/pricing/src/*"],
       "@genfeedai/hooks/*": ["../../packages/hooks/*"],
       "@genfeedai/next-config": [
         "../../packages/next-config/next.config.base.ts"
