+++ apps/server/api/vitest.config.ts
@@ -140,6 +140,10 @@ export default defineConfig({
         replacement: path.resolve(serviceDir, '../../../packages/config/src'),
       },
       {
+        find: '@genfeedai/pricing',
+        replacement: path.resolve(serviceDir, '../../../packages/pricing/src'),
+      },
+      {
         find: '@genfeedai/harness',
         replacement: path.resolve(serviceDir, '../../../packages/harness/src'),
       },
