+++ packages/client/tsconfig.json
@@ -19,6 +19,8 @@
       "@genfeedai/enums/*": ["../enums/src/*"],
       "@genfeedai/helpers": ["../helpers/src/index.ts"],
       "@genfeedai/helpers/*": ["../helpers/src/*"],
+      "@genfeedai/pricing": ["../pricing/src/index.ts"],
+      "@genfeedai/pricing/*": ["../pricing/src/*"],
       "@genfeedai/interfaces": ["../interfaces/src/index.ts"],
       "@genfeedai/interfaces/*": ["../interfaces/src/*"],
       "@helpers/*": ["../helpers/src/*"],
