+++ packages/serializers/tsconfig.json
@@ -11,6 +11,8 @@
       "@genfeedai/enums/*": ["../enums/src/*"],
       "@genfeedai/helpers": ["../helpers/src/index.ts"],
       "@genfeedai/helpers/*": ["../helpers/src/*"],
+      "@genfeedai/pricing": ["../pricing/src/index.ts"],
+      "@genfeedai/pricing/*": ["../pricing/src/*"],
       "@genfeedai/serializers/*": ["./src/*"],
       "@serializers/*": ["./src/*"]
     },
